fix(core): move lottieReadiness out of src/runtime so it ships in dist
The "./runtime/lottie-readiness" subpath export claimed to point at
./dist/runtime/adapters/lottieReadiness.js, but the published 0.6.6 and
0.6.7 tarballs never contained that file. Reason: packages/core/tsconfig.json
excludes "src/runtime" (those files run in a browser context and are
bundled separately into the IIFE artifact), so tsc never emitted the
compiled output. Consumers that import the subpath — most notably
@hyperframes/studio's Player.tsx — fail to resolve the module and the
build breaks at the consumer site.

lottieReadiness.ts is a pure helper: takes `unknown`, returns `boolean`,
no DOM or `window` access. It doesn't belong under src/runtime/ in the
first place — that's why the runtime-exclude rule rightly caught it.
Move it to src/lottieReadiness.ts so the standard library build picks
it up. The subpath export name stays "./runtime/lottie-readiness" so
existing consumers (studio) don't need a code change; only the exports
map's underlying file path moves to ./dist/lottieReadiness.{js,d.ts}.

- mv src/runtime/adapters/lottieReadiness.{ts,test.ts} -> src/
- update src/runtime/adapters/lottie.ts re-export path
- update package.json + publishConfig.exports to point at new dist path

Verified: full core test suite passes (862/862), dist now contains
lottieReadiness.{js,d.ts}, studio typechecks against the moved file.
