+++ packages/core/src/lottieReadiness.test.ts
