+++ packages/core/src/lottieReadiness.ts
