+++ packages/core/src/runtime/adapters/lottie.ts
@@ -1,6 +1,6 @@
 import type { RuntimeDeterministicAdapter } from "../types";
 import { swallow } from "../diagnostics";
-export { isLottieAnimationLoaded } from "./lottieReadiness";
+export { isLottieAnimationLoaded } from "../../lottieReadiness";
 
 /**
  * Lottie adapter for HyperFrames
